+++ packages/jest/index.js
@@ -2,5 +2,12 @@ import {nativeConfig} from './native/config.js'
 import {webConfig} from './web/config.js'
 
 export default {
-  projects: [webConfig, nativeConfig]
+  projects: [webConfig, nativeConfig],
+  collectCoverageFrom: [
+    '**/*.{ts,tsx}',
+    '!**/*.test.*',
+    '!**/*.d.ts',
+    '!**/node_modules/**',
+    '!**/dist/**'
+  ]
 }