+++ packages/primitives/src/primitive/Box/Box.tsx
@@ -1,7 +1,5 @@
-// TODO: When extending via ComponentPropsWithoutRef we lose
-// the children prop
 import {getEnvironmentName} from '@interunit/config'
-import {type Merge} from '@interunit/toolbox'
+import {type MergedCrossPlatformProps} from '@interunit/toolbox'
 import React from 'react'
 import type {View} from 'react-native'
 
@@ -28,19 +26,7 @@ const Box = React.forwardRef(
       el,
       children,
       ..._props
-    }: BoxPrimitiveProps<T> & {
-      style?: Merge<
-        [
-          React.ComponentPropsWithoutRef<typeof View>,
-          React.JSX.IntrinsicElements[T]['style']
-        ]
-      >
-    } & Merge<
-        [
-          React.ComponentPropsWithoutRef<typeof View>,
-          React.JSX.IntrinsicElements[T]
-        ]
-      >,
+    }: BoxPrimitiveProps<T> & MergedCrossPlatformProps<T, typeof View>,
     forwardedRef: any
   ) => {
     // TODO: Why does this need to be re-casted to work
@@ -82,6 +68,8 @@ const Box = React.forwardRef(
       </Box>
     )
   }
-)
+) as <T extends keyof typeof BoxConstruct>(
+  props: BoxPrimitiveProps<T> & MergedCrossPlatformProps<T, typeof View>
+) => React.JSX.Element
 
 export {Box}