+++ packages/prettier/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @interunit/prettier
 
+## 0.0.3
+
+### Patch Changes
+
+- Workspaces update + Types fixes
+
 ## 0.0.2
 
 ### Patch Changes