+++ packages/tabs/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @interunit/tabs
 
+## 0.0.9
+
+### Patch Changes
+
+- Workspaces update + Types fixes
+- Updated dependencies
+  - @interunit/primitives@0.0.17
+  - @interunit/toolbox@0.0.11
+  - @interunit/config@0.0.12
+  - @interunit/a11y@0.0.15
+
 ## 0.0.8
 
 ### Patch Changes